renamed test file to module instead of function
