--- server/test/queryBuilder.test.js
@@ -1,50 +1,0 @@
-/* eslint no-unused-vars: 0 */
-const jest = require('jest');
-const { queryBuilder } = require('../lib/sightingsController.js');
-
-describe('queryBuilder', () => {
-	test('should return an empty JSON when no inputs', () => {
-		const searchOpts = {};
-		expect(queryBuilder(searchOpts)).toEqual({});
-	});
-
-	test('Should add user id correctly as a search option', () => {
-		const searchOpts = {
-			id: 'newuser'
-		};
-		expect(queryBuilder(searchOpts)).toEqual({
-			spottedBy: { $in: 'newuser' }
-		});
-	});
-
-	test('Should add location correctly as a search option', () => {
-		const searchOpts = {
-			location: 'home'
-		};
-		expect(queryBuilder(searchOpts)).toEqual({
-			location: { $in: 'home' }
-		});
-	});
-
-	test('Should add animal correctly as a search option', () => {
-		const searchOpts = {
-			animal: 'cat'
-		};
-		expect(queryBuilder(searchOpts)).toEqual({
-			animal: { $in: 'cat' }
-		});
-	});
-
-	test('Should correctly add all inputs as search options', () => {
-		const searchOpts = {
-			id: 'newuser',
-			animal: 'cat',
-			location: 'home',
-		};
-		expect(queryBuilder(searchOpts)).toEqual({
-			spottedBy: { $in: 'newuser' },
-			animal: { $in: 'cat' },
-			location: { $in: 'home' }
-		});
-	});
-});
